Common getTokenType used on UI, monitor and Oracle (#165)
* Stopped silently swallowing errors

* getTokenType in commons

* Oracle using getTokenType from commons

* Moved the getTokenType tests from ui to commons

* Using getTokenType in ui and oracle

* Revert "Stopped silently swallowing errors"

fc92c0af9dc7cb4e59a7f1779afc8da8d9c33ecf

* using common getTokenType in monitor.

* Lint.
