+++ monitor/utils/events.js
@@ -8,9 +8,10 @@ const {
   getBridgeABIs,
   getBridgeMode,
   HOME_ERC_TO_ERC_ABI,
-  ERC20_ABI
+  ERC20_ABI,
+  ERC677_BRIDGE_TOKEN_ABI,
+  getTokenType
 } = require('../../commons')
-const { getTokenType } = require('./ercUtils')
 
 const { HOME_RPC_URL, FOREIGN_RPC_URL, HOME_BRIDGE_ADDRESS, FOREIGN_BRIDGE_ADDRESS } = process.env
 const HOME_DEPLOYMENT_BLOCK = toBN(Number(process.env.HOME_DEPLOYMENT_BLOCK) || 0)
@@ -30,12 +31,15 @@ async function main(mode) {
   const { HOME_ABI, FOREIGN_ABI } = getBridgeABIs(bridgeMode)
   const homeBridge = new web3Home.eth.Contract(HOME_ABI, HOME_BRIDGE_ADDRESS)
   const foreignBridge = new web3Foreign.eth.Contract(FOREIGN_ABI, FOREIGN_BRIDGE_ADDRESS)
-  const tokenType = await getTokenType(foreignBridge, FOREIGN_BRIDGE_ADDRESS)
-  const isExternalErc20 = tokenType === ERC_TYPES.ERC20
   const v1Bridge = bridgeMode === BRIDGE_MODES.NATIVE_TO_ERC_V1
   const erc20MethodName =
     bridgeMode === BRIDGE_MODES.NATIVE_TO_ERC || v1Bridge ? 'erc677token' : 'erc20token'
   const erc20Address = await foreignBridge.methods[erc20MethodName]().call()
+  const tokenType = await getTokenType(
+    new web3Foreign.eth.Contract(ERC677_BRIDGE_TOKEN_ABI, erc20Address),
+    FOREIGN_BRIDGE_ADDRESS
+  )
+  const isExternalErc20 = tokenType === ERC_TYPES.ERC20
   const erc20Contract = new web3Foreign.eth.Contract(ERC20_ABI, erc20Address)
 
   logger.debug('getting last block numbers')
