+++ oracle/scripts/initialChecks.js
@@ -1,6 +1,6 @@
 require('../env')
 const Web3 = require('web3')
-const { ERC677_BRIDGE_TOKEN_ABI, ERC_TYPES } = require('../../commons')
+const { ERC677_BRIDGE_TOKEN_ABI, getTokenType } = require('../../commons')
 
 async function initialChecks() {
   const { ERC20_TOKEN_ADDRESS, BRIDGE_MODE, FOREIGN_RPC_URL, FOREIGN_BRIDGE_ADDRESS } = process.env
@@ -8,17 +8,12 @@ async function initialChecks() {
 
   if (BRIDGE_MODE === 'ERC_TO_ERC') {
     const foreignWeb3 = new Web3(new Web3.providers.HttpProvider(FOREIGN_RPC_URL))
-    const tokenContract = new foreignWeb3.eth.Contract(ERC677_BRIDGE_TOKEN_ABI, ERC20_TOKEN_ADDRESS)
-    try {
-      const bridgeContract = await tokenContract.methods.bridgeContract().call()
-      if (bridgeContract === FOREIGN_BRIDGE_ADDRESS) {
-        result.foreignERC = ERC_TYPES.ERC677
-      } else {
-        result.foreignERC = ERC_TYPES.ERC20
-      }
-    } catch (e) {
-      result.foreignERC = ERC_TYPES.ERC20
-    }
+    const bridgeTokenContract = new foreignWeb3.eth.Contract(
+      ERC677_BRIDGE_TOKEN_ABI,
+      ERC20_TOKEN_ADDRESS
+    )
+
+    result.foreignERC = await getTokenType(bridgeTokenContract, FOREIGN_BRIDGE_ADDRESS)
   }
   console.log(JSON.stringify(result))
   return result
