+++ ui/src/stores/utils/contract.js
@@ -157,19 +157,6 @@ export const getDeployedAtBlock = async contract => {
   }
 }
 
-export const getTokenType = async (contract, bridgeAddress) => {
-  try {
-    const bridgeContract = await contract.methods.bridgeContract().call()
-    if (bridgeContract === bridgeAddress) {
-      return ERC_TYPES.ERC677
-    } else {
-      return ERC_TYPES.ERC20
-    }
-  } catch (e) {
-    return ERC_TYPES.ERC20
-  }
-}
-
 export const getBlockRewardContract = contract => contract.methods.blockRewardContract().call()
 
 export const getValidatorContract = contract => contract.methods.validatorContract().call()
